discord bot/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.1.dev0
discord.py==2.0
fastapi==0.111.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
pydantic==2.7.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
yarl==1.9.4